add travis.yml; move instruction into readme; clear environment.yml; add setup.cfg; move all code inside 'sound-recognition' folder
