+++ .travis.yml
@@ -1,0 +1,35 @@
+os: linux
+
+dist: bionic
+
+language: python
+
+python:
+  - "3.7"
+
+sudo: required
+
+before_install:
+  - sudo apt-get update
+  - wget https://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh -O miniconda.sh;
+  - bash miniconda.sh -b -p $HOME/miniconda
+  - source "$HOME/miniconda/etc/profile.d/conda.sh"
+  - hash -r
+  - conda config --set always_yes yes --set changeps1 no
+  - conda update -q conda
+  # Useful for debugging any issues with conda
+  - conda info -a
+
+install:
+  - conda env create -f environment.yml
+  - conda activate sound-recog
+
+
+before_script:
+  - python --version
+  - conda list
+
+script:
+  - black sound-recognition
+  - isort -y -rc sound-recognition
+  - flake8 sound-recognition
