--- Instruction
@@ -1,10 +1,0 @@
-"""
-conda create -n pack_area -f environment.yml
-conda activate pack_area
-
-# start server:
-python app.py
-
-# client part for recording wav files and sending it to server
-python request.py
-"""
