+++ environment.yml
@@ -1,36 +1,11 @@
-name: pack_area
+name: sound-recog
 channels:
-  - defaults
+  - conda-forge
 dependencies:
-  - asn1crypto=1.3.0=py37_0
-  - ca-certificates=2020.1.1=0
-  - certifi=2019.11.28=py37_0
-  - cffi=1.14.0=py37h7a1dbc1_0
-  - chardet=3.0.4=py37_1003
-  - click=7.0=py37_0
-  - cryptography=2.8=py37h7a1dbc1_0
-  - flask=1.1.1=py_0
-  - idna=2.8=py37_0
-  - itsdangerous=1.1.0=py37_0
-  - jinja2=2.11.1=py_0
-  - markupsafe=1.1.1=py37he774522_0
-  - openssl=1.1.1d=he774522_4
-  - pip=20.0.2=py37_1
-  - portaudio=19.6.0=he774522_4
-  - pyaudio=0.2.11=py37hfa6e2cd_1
-  - pycparser=2.19=py37_0
-  - pyopenssl=19.1.0=py37_0
-  - pysocks=1.7.1=py37_0
-  - python=3.7.6=h60c2a47_2
-  - requests=2.23.0=py37_0
-  - setuptools=46.0.0=py37_0
-  - six=1.14.0=py37_0
-  - sqlite=3.31.1=he774522_0
-  - urllib3=1.25.8=py37_0
-  - vc=14.1=h0510ff6_4
-  - vs2015_runtime=14.16.27012=hf0eaf9b_1
-  - werkzeug=1.0.0=py_0
-  - wheel=0.34.2=py37_0
-  - win_inet_pton=1.1.0=py37_0
-  - wincertstore=0.2=py37_0
-
+  - flask>=1.1.1
+  - pyaudio>=0.2.11
+  - python=3.7.6
+  - requests>=2.23.0
+  - flake8
+  - black
+  - isort
