+++ setup.cfg
@@ -1,0 +1,3 @@
+[flake8]
+ignore = D202,D203,W503,E203
+max-line-length = 119
