+++ sound-recognition/Client-Server/static/css/home.css
